# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vensle API ",
        "description": "Description of your API",
        "contact": {
            "email": "hybiekay2@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/register": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Register a new user",
                "operationId": "b41b30d082ecc16fbebf247bea1125a2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "phone_number",
                                    "address"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Street, City"
                                    },
                                    "business_name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Acme Inc. ",
                                        "nullable": true
                                    },
                                    "business_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "info@acme.com",
                                        "nullable": true
                                    },
                                    "phone": {
                                        "type": "nullable",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "business_address": {
                                        "type": "string",
                                        "example": "456 Business Street, Business City",
                                        "nullable": true
                                    },
                                    "certificate": {
                                        "type": "string",
                                        "example": "Certificate of Business",
                                        "nullable": true
                                    },
                                    "bank_name": {
                                        "type": "string",
                                        "example": "Bank of Business",
                                        "nullable": true
                                    },
                                    "account_number": {
                                        "type": "string",
                                        "example": "1234567890",
                                        "nullable": true
                                    },
                                    "profile_picture": {
                                        "type": "string",
                                        "example": "https://example.com/profile.jpg",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful registration"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/v1/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login user",
                "operationId": "2f6428dad59f777ed98e70646173d7d7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": "3600"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/v1/update-profile": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user",
                "operationId": "06340a99c2d14685140054f07635bd1b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe",
                                        "nullable": true
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com",
                                        "nullable": true
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "1234567890",
                                        "nullable": true
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Street, City",
                                        "nullable": true
                                    },
                                    "profile_picture": {
                                        "type": "file",
                                        "example": null,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get current user details",
                "operationId": "8031868a75a54217ac0c23a3cbefe117",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email"
                                        },
                                        "phone_number": {
                                            "type": "string"
                                        },
                                        "address": {
                                            "type": "string"
                                        },
                                        "profile_picture": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/update-passwords": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user's password",
                "operationId": "cef1f44c64495db1d14b1ac8ad2eff6a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "old_password",
                                    "new_password",
                                    "new_password_confirmation"
                                ],
                                "properties": {
                                    "old_password": {
                                        "type": "string"
                                    },
                                    "new_password": {
                                        "type": "string"
                                    },
                                    "new_password_confirmation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}